+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.9.0 (unreleased)
+## 0.9.0 (2024-10-17)
 
 - Updated XGBoost to 2.1.1
 - Added support for callbacks